+++ core/src/main/kotlin/io/github/serpro69/kfaker/provider/Phish.kt
@@ -12,5 +12,14 @@ class Phish internal constructor(fakerService: FakerService) : AbstractFakeDataP
     override val localUniqueDataProvider = LocalUniqueDataProvider<Phish>()
     override val unique by UniqueProviderDelegate(localUniqueDataProvider)
 
-    fun song() = resolve("song")
+    @Deprecated(
+        message = "Deprecated since 1.5.0 due to changes in dict file. Will be removed in 1.6.0",
+        replaceWith = ReplaceWith("songs()"),
+        level = DeprecationLevel.WARNING
+    )
+    fun song() = songs()
+
+    fun albums() = resolve("albums")
+    fun musicians() = resolve("musicians")
+    fun songs() = resolve("songs")
 }